requirements.txt
streamlit
google-generativeai
yt-dlp
requests